refactor(trusty-search): extract shared confirm() helper; prune-orphans uses it
Extract the y/N confirmation prompt from prune_orphans.rs into a
pub(crate) fn in commands/mod.rs so both prune commands share one
implementation. prune_orphans.rs now calls super::confirm().

This is the infrastructure change preceding the prune.rs safety fixes.
